Matching data types to internal PySpark and pandas' behavior. (#1870)
This PR basically includes several fixes for matching data types to internal PySpark and pandas' behavior :
- Added `np.float32` and `"float32"` (matched to `FloatType`)
    ```python
    >>> ks.Series([10]).astype(np.float32)
    0    10.0
    dtype: float32

    >>> ks.Series([10]).astype("float32")
    0    10.0
    dtype: float32
    ```
- Added `np.datetime64` and `"datetime64[ns]"` (matched to `TimestampType`) 
    ```python
    >>> ks.Series(["2020-10-26"]).astype(np.datetime64)
    0   2020-10-26
    dtype: datetime64[ns]

    >>> ks.Series(["2020-10-26"]).astype("datetime64[ns]")
    0   2020-10-26
    dtype: datetime64[ns]
    ```
- Fixed `np.int` to match `LongType`, not `IntegerType`.
    ```python
    >>> pd.Series([100]).astype(np.int)
    0    100.0
    dtype: int64

    >>> ks.Series([100]).astype(np.int)
    0    100.0
    dtype: int32  # This fixed to `int64` now.
    ```
- Fixed `np.float` to match `DoubleType`, not `FloatType`.
    ```python
    >>> pd.Series([100]).astype(np.float)
    0    100.0
    dtype: float64

    >>> ks.Series([100]).astype(np.float)
    0    100.0
    dtype: float32  # This fixed to `float64` now.
    ```
- Added test, `test_typedef.py::TypeHintTests::test_as_spark_type`
- Added more tests for `test_series.py::SeriesTest::test_astype`
- Fixed related doc tests

+++ databricks/koalas/accessors.py
@@ -489,7 +489,7 @@ class KoalasFrameMethods(object):
         0    3
         1    5
         2    7
-        dtype: int32
+        dtype: int64
 
         You can also omit the type hints so Koalas infers the return schema as below:
